corrections entitées base de données & début websocket

+++ bootcamp-helpcenter/src/main/java/com/example/necroliner/bootcamphelpcenter/model/Conversation.java
@@ -1,29 +1,25 @@
 package com.example.necroliner.bootcamphelpcenter.model;
-import java.util.ArrayList;
-import java.util.List;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToMany;
 
 @Entity
 public class Conversation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+
     private boolean archive;
+    
     public Conversation() {
         this.archive=false;
     }
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
     public boolean isArchive() {
